Prefer using retinex plugin for retinex mask
I discovered that when this function takes the custom expr path,
i.e. when it has akarin and vszip available, the speed is an
order of magnitude slower than when taking the `retinex.MSRCP` path.

Further investigation revealed this is a result of `gauss_blur` and
the high `sigma` values being used here. `gauss_blur` has been discussed
previously as a problematic function, and we don't currently have an
alternative which maintains the same quality, so my current recommendation
is to avoid this path for the time being.